requirements.txt
torch
torchinfo
loguru
numpy
scikit-learn
matplotlib
seaborn
tqdm
mlflow
pyyaml
dvc[gdrive]